Call external service with RestClient

+++ src/main/java/com/workbeattalent/sport_club/user/UserRestClient.java
@@ -25,7 +25,7 @@ public class UserRestClient {
                 });
     }
 
-    public User getById(final Long id) {
+    public User getById(final Integer id) {
         return this.restClient.get()
                 .uri("/users/{id}", id)
                 .retrieve()
